third_party/blink/web_tests/editing/shadow/compare-positions-in-nested-shadow-expected.txt

PASS selection2.anchorNode is selection2.focusNode
PASS selection2.anchorOffset is 5
PASS selection2.focusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
